+++ Sources/Target/FileTarget.swift
@@ -25,6 +25,7 @@ public class FileTarget: Target {
     }
     
     public func send(_ level: LogLevel, _ message: String, _ metaInfo: MetaInfo) {
+        guard self.config.levels.contains(level) else { return }
         self.write("\(metaInfo.file) \(metaInfo.function) \(metaInfo.line) \(level.title): \(message)\n")
     }